made changes according to PR comments
