+++ src/transform_data.py
@@ -3,7 +3,6 @@ Module containing functions for transforming data
 to be put into deep learning model
 """
 
-
 import numpy as np
 from skimage import transform
 from fastai.vision import Image
